requirements.txt
# MCP и связанные пакеты
mcp>=0.1.0

# Web framework
starlette>=0.27.0
uvicorn>=0.24.0

# OAuth2 и безопасность
pyjwt>=2.8.0
cryptography>=41.0.0

# Database
aiosqlite>=0.19.0  # Для SQLite event store

# Типизация для старых версий Python
typing-extensions>=4.8.0

# Асинхронность
anyio>=4.0.0

# CLI
click>=8.1.0

# Утилиты
python-multipart>=0.0.6  # Для обработки форм в Starlette
pytz>=2023.3  # Для работы с временными зонами

# Тестирование
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # Для TestClient в Starlette

# Для продакшена (опционально)
# redis>=5.0.0  # Для хранения сессий
# bcrypt>=4.1.0  # Для безопасного хеширования паролей
# gunicorn>=21.0.0  # Production ASGI server
# sqlalchemy>=2.0.0  # Для хранения данных в БД
# alembic>=1.12.0  # Для миграций БД
# sentry-sdk>=1.38.0  # Для мониторинга ошибок
numpy==2.1.3
openai==1.57.0
aiosqlite==0.20.0
